+++ packages/implementations/react/src/atomic/Progress/src/Progress.tsx
@@ -7,15 +7,15 @@ const Progress: React.FC<ProgressProps> & ProgressComponents = ({
   className: _className,
   style: _style,
   percentage,
-  backgroundColor = "$info",
+  color = "$info",
   ...props
 }: ProgressProps) => {
   const dataTestid = (props as { "data-testid": string })?.["data-testid"];
   const completed = percentage >= 100;
 
-  const { className, style, otherProps } = progress.sprinkle({
+  const { className, style } = progress.sprinkle({
     width: `${percentage}%`,
-    backgroundColor: completed ? "$success" : backgroundColor,
+    backgroundColor: completed ? "$success" : color,
   });
 
   return (
@@ -23,7 +23,6 @@ const Progress: React.FC<ProgressProps> & ProgressComponents = ({
       <div
         className={[progress.classnames.progressBar, className].join(" ")}
         style={style}
-        {...otherProps}
         data-testid={dataTestid ? `${dataTestid}-bar` : ""}
       />
     </div>