+++ packages/implementations/react/src/atomic/Progress/src/components/Circle/Circle.tsx
@@ -7,11 +7,16 @@ const ProgressCircle: React.FC<ProgressCircleProps> = ({
   className: _className,
   style: _style,
   percentage,
+  color = "$info",
   size = "large",
   ...props
 }) => {
   const completed = percentage >= 100;
   const dataTestid = (props as { "data-testid": string })?.["data-testid"];
+
+  const { className, style } = progress.sprinkle({
+    stroke: completed ? "$success" : color,
+  });
 
   return (
     <div className={progress.classnames.circle({ size })} {...props}>
@@ -26,9 +31,8 @@ const ProgressCircle: React.FC<ProgressCircleProps> = ({
           strokeLinecap="round"
         />
         <circle
-          className={progress.classnames.circleProgress({
-            completed,
-          })}
+          className={[progress.classnames.circleProgress, className].join(" ")}
+          style={style}
           cx="70"
           cy="70"
           strokeWidth={percentage === 0 ? 0 : 8}