+++ packages/implementations/react/src/atomic/Progress/src/components/Circle/circle.spec.tsx
@@ -4,32 +4,43 @@ import { render, screen } from "@testing-library/react";
 import { ProgressCircle } from "./Circle";
 import { ProgressCircleProps } from "./circle.types";
 
-const makeSut = ({ percentage, size }: ProgressCircleProps) => {
+const makeSut = ({ percentage, size, color }: ProgressCircleProps) => {
   render(
     <ProgressCircle
       percentage={percentage}
       size={size}
+      color={color}
       data-testid="progress-element"
     />,
   );
 };
+
 describe("GIVEN <Progress.Cicle />", () => {
   describe("WHEN rendered", () => {
     it("THEN should display the correct percentage", () => {
       makeSut({ percentage: 50 });
       expect(screen.getByText("50%")).toBeDefined();
     });
+
     it("THEN should display the corrected style when not complete", () => {
       makeSut({ percentage: 50 });
       expect(
         screen.getByTestId("progress-element-bar").getAttribute("class"),
-      ).not.toContain("completed_true");
+      ).not.toContain("stroke-success");
     });
+
     it("THEN should display the corrected style when complete", () => {
       makeSut({ percentage: 100 });
       expect(
         screen.getByTestId("progress-element-bar").getAttribute("class"),
-      ).toContain("completed_true");
+      ).toContain("stroke-success");
+    });
+
+    it("THEN should display the corrected style when brand", () => {
+      makeSut({ percentage: 50, color: "$brand" });
+      expect(
+        screen.getByTestId("progress-element-bar").getAttribute("class"),
+      ).toContain("stroke-brand");
     });
   });
 
@@ -40,6 +51,7 @@ describe("GIVEN <Progress.Cicle />", () => {
         screen.getByTestId("progress-element").getAttribute("class"),
       ).toContain("size_large");
     });
+
     it("THEN should correctly render the size small", () => {
       makeSut({ percentage: 50, size: "small" });
       expect(