+++ packages/implementations/react/src/atomic/Textarea/src/Textarea.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { textarea } from "@rarui/styles";
+
+import { TextareaProps } from "./textarea.types";
+
+const Textarea: React.FC<TextareaProps> = ({
+  className: _className,
+  style: _style,
+  lines = 2,
+  appearance,
+  ...rest
+}) => (
+  <textarea
+    className={textarea.classnames.textarea({ appearance })}
+    rows={lines}
+    {...rest}
+  />
+);
+
+Textarea.displayName = "Textarea";
+
+export { Textarea };
+export default Textarea;