+++ packages/implementations/react/src/atomic/Textarea/webpack.config.ts
@@ -1,0 +1,11 @@
+import path from "path";
+import { configuration } from "@rarui/webpack";
+
+const config = {
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    library: "@rarui-react/textarea",
+  },
+};
+
+export default () => configuration.getConfiguration(config);