+++ packages/styles/src/packages/atomic/textarea/rarui-textarea.types.ts
@@ -1,0 +1,6 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { textareaStyles } from "./rarui-textarea.css";
+
+export type TextareaVariants = NonNullable<
+  RecipeVariants<typeof textareaStyles.textarea>
+>;