+++ packages/implementations/react/src/atomic/Status/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Status are used to indicate the current state of a process.
 
+## 2024-07-04 `1.1.0`
+
+#### 🎉 New features
+
+- Added new `dot` and `full` properties to the component. ([#71](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/71) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-06-02 `1.0.2`
 
 #### 💡 Others