+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 RarUI Styles deprive all styles needed to build components.
 
+## 2024-07-04 `1.32.0`
+
+#### 🎉 New features
+
+- Added new style pack for `textarea` component. ([#71](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/71) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new `full` property to the component `Status`. ([#71](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/71) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-07-03 `1.31.1`
 
 #### 🎉 New features